Remove legacy ThemeContext from DOM exports
The theme context is a legacy API that was only used on native to set
theme variables. We now have a native shim for the StyleX theme APIs,
and no longer require the theme context. However, internally Meta still
uses ThemeContext for native as interal StyleX use has not migrated to
StyleX theme APIs.

This patch removes ThemeContext from the DOM exports, which allows the
library to be used in React Server Components.

Close #76